server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.3
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
alembic==1.13.2
pydantic==2.8.2
PyJWT==2.9.0
cryptography>=42.0.0
passlib[bcrypt]==1.7.4
redis==5.0.8
prometheus-client==0.20.0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytest==8.3.2
httpx==0.27.2
pytest-asyncio==0.23.8
black==24.8.0
ruff==0.6.9
mypy==1.11.2
requests==2.32.3
openapi-spec-validator==0.7.1

